pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gtm"
version = "0.1.0"
description = "GTM-as-Code: Python CLI + MCP server for Google Tag Manager management"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "pydantic>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "fastmcp>=2.0",
    "deepdiff>=7.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
gtm = "gtm.cli:app"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
